requirements.txt
Django==2.2.16
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-dotenv==0.19.0